+++ robot_descriptions/_descriptions.py
@@ -105,6 +105,7 @@ DESCRIPTIONS: Dict[str, Description] = {
     "laikago_description": Description(Format.URDF),
     "mini_cheetah_description": Description(Format.URDF),
     "minitaur_description": Description(Format.URDF),
+    "nextage_description": Description(Format.URDF),
     "panda_description": Description(Format.URDF),
     "panda_mj_description": Description(Format.MJCF),
     "pepper_description": Description(Format.URDF),